refactor: 패키지 경로 롤백
